Move record building logic out of WARCWriter.
This make it easier creating record without having an handle to a
writer.
